Add support for parsing of !Type:Invst transactions
Add InvestmentTransaction interface/type and a parsing function,
matching how this was implemented for BankingTransaction.

Read() is updated to parse Banking or Investment transactions
based on Header.
As Read() is returning type Transaction, extend the interface
with a TransactionType() so that user can do the appriopriate
typecast to either BankTransaction or InvestmentTransaction.

v2: Fix Investment fields which were swapped (Price and Shares)